requirements.txt
PyYAML==3.10
Pygments==1.6
argparse==1.2.1
coverage==3.6
distribute==0.6.24
fancycompleter==0.4
ipython==0.13.2
lxml==3.2.1
nltk==2.0.4
nose==1.3.0
numpy==1.7.1
pandas==0.11.0
pdbpp==0.7.2
pyrepl==0.8.4
python-dateutil==2.1
pytz==2013b
scikit-learn==0.13.1
six==1.3.0
ujson==1.30
unicodecsv==0.9.4
winpdb==1.4.8
wmctrl==0.1
wsgiref==0.1.2
